apps/ai-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic>=2.9.0,<3.0.0
httpx>=0.27.0,<1.0.0
langchain>=0.3.0,<0.4.0
langchain-community>=0.3.0,<0.4.0
langgraph>=0.2.28,<0.3.0
sentence-transformers>=3.1.0,<4.0.0
openai>=1.50.0,<2.0.0
python-dotenv>=1.0.1,<2.0.0
numpy>=1.26.4,<2.0.0
openai-whisper>=20231117
